refactor: nested file structure & fix: changing shooting key from enter to space
