+++ index.html
@@ -14,7 +14,7 @@
     <div id="instructions">
       <p>Instructions:</p>
       <p>- Use the arrow keys to move the tank around</p>
-      <p>- Press enter to shoot</p>
+      <p>- Press Space to shoot</p>
     </div>
     <canvas id="canvas">
     </canvas>
@@ -38,7 +38,7 @@
   <img id="fire10" src="./pictures/tank_explosion10.png">
   <img id="fire11" src="./pictures/tank_explosion11.png">
   <img id="fire12" src="./pictures/tank_explosion12.png">
-  <script type="module" src="/main.js"></script>
+  <script type="module" src="/src/main.js"></script>
 </body>
 
 </html>
