+++ package.json
@@ -12,4 +12,3 @@
     "vite": "^4.0.0"
   }
 }
-
