+++ src/logic/animation/circle.js
