+++ src/logic/animation/particle.js
