+++ src/logic/input/input.js
@@ -13,29 +13,33 @@ export class InputHandler {
         e.key === "ArrowDown" ||
         e.key === "ArrowUp" ||
         e.key === "ArrowLeft" ||
-        e.key === "Enter" ||
-        (e.key === "ArrowRight" && this.keys.indexOf(e.key) === -1)
+        e.key === " " || // Space key
+        e.key === "ArrowRight"
       ) {
+        if (this.keys.indexOf(e.key) === -1) {
+          this.keys.push(e.key);
+        }
+
         this.keyPress = true;
         this.lastKey = e.key;
         this.player.update(this.lastKey);
+
         this.circles.forEach((circle, index) => {
-          if (e.key === "Enter" && index === this.circles.length - 1) {
+          if (e.key === " " && index === this.circles.length - 1) {
             const { x, y, angle } = this.tank.getTurretPosition();
             circle.releaseCircle(x, y, angle);
-          }
-          if (
+          } else if (
             e.key === "ArrowLeft" ||
             (e.key === "ArrowRight" && index === this.circles.length - 1)
           ) {
             circle.moveCircle(e.key);
-          }
-          if (e.key === "ArrowUp" || e.key === "ArrowDown") {
+          } else if (e.key === "ArrowUp" || e.key === "ArrowDown") {
             const { x, y, angle } = this.tank.getTurretPosition();
             circle.moveCircle(e.key, x, y, angle);
           }
         });
-        if (e.key === "Enter" && this.game.gameOver) {
+
+        if (e.key === " " && this.game.gameOver) {
           this.game.restart();
           this.game.start();
         }
@@ -45,6 +49,10 @@ export class InputHandler {
     window.addEventListener("keyup", (e) => {
       this.keyPress = false;
       this.lastKey = "";
+      const index = this.keys.indexOf(e.key);
+      if (index !== -1) {
+        this.keys.splice(index, 1);
+      }
     });
   }
 }
